requirements.txt
# Python Dependencies
feedparser==6.0.10
requests==2.31.0
google-generativeai==0.8.3
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
